+++ src/main/java/com/electronics/exception/MerchantNotFoundException.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 public class MerchantNotFoundException extends EcommerceException {
 
-    public MerchantNotFoundException(Long id) {
+    public MerchantNotFoundException(Integer id) {
         super("Merchant with id=%s not found".formatted(id), HttpStatus.NOT_FOUND.value(),
                 "MERCHANT_NOT_FOUND", Map.of("id", id));
     }
