+++ src/main/java/com/electronics/repository/MerchantRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MerchantRepository extends JpaRepository<Merchant, Long> {
+public interface MerchantRepository extends JpaRepository<Merchant, Integer> {
 
     Optional<Merchant> findByUsername(String username);
 }
